deleted unnessary javscript files and moved word problems into folder
