+++ src/wordproblems/wordproblems1.ts
@@ -1,7 +1,7 @@
 //word problem one
 //What are all of the words containing UU?
 
-import { tempWords } from "../scripts/tempWordList";
+import { tempWords } from "../../scripts/tempWordList";
 // console.log(tempWords);
 
 function returnsWordsWithUUIncluded(arr: string[]): string[] {
