+++ src/wordproblems/wordproblems10.ts
