+++ src/wordproblems/wordproblems4.ts
@@ -1,4 +1,4 @@
-import { tempWords } from "../scripts/tempWordList";
+import { tempWords } from "../../scripts/tempWordList";
 
 // What are all of the words that contain the word CAT and are exactly 5 letters long?
