+++ src/wordproblems/wordproblems5.ts
@@ -1,4 +1,4 @@
-import { tempWords } from "../scripts/tempWordList";
+import { tempWords } from "../../scripts/tempWordList";
 
 // What are all of the words that have no E or A and are at least 15 letters long?
